project-root/spark-job/requirements.txt
pyspark
python-dotenv

